+++ mailscanner/mcp_rules_update.php
@@ -43,7 +43,7 @@ if ($_SESSION['user_type'] != 'A') {
 } else {
     html_start("MCP Rule Description Update", 0, false, false);
 
-    echo '<form method="post" action="' . $_SERVER['PHP_SELF'] . '" >' . "\n";
+    echo '<form method="post" action="' . sanitizeInput($_SERVER['PHP_SELF']) . '" >' . "\n";
     echo '<input type="hidden" name="run" value="true">' . "\n";
     echo '<table class="boxtable" width="100%">' . "\n";
     echo ' <tr>' . "\n";