+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 -----------
  - Correct unexpected behaviour in viewpart.php if one or more headers are not set
  - Refactor message part viewer
+ - Sanitize user input
 
 1.2.0 - Beta 7
 -----------